Remove old runner model and helpers.rs
