+++ sp-runner/src/formal_model/transition_system_model.rs
@@ -1,6 +1,12 @@
 /// This module contain a simple model type that we can use for the
 /// formal verification stuff.
+use serde::{Deserialize, Serialize};
+use sp_domain::*;
 use super::*;
+
+// remember to also change in planner
+// TODO: make this constant crate-wide and use in more places
+const USE_GUARD_EXTRACTION: bool = false;
 
 #[derive(Debug, PartialEq, Serialize, Deserialize, Clone, Default)]
 pub struct TransitionSystemModel {
@@ -140,13 +146,56 @@ impl TransitionSystemModel {
             });
         }
 
-        TransitionSystemModel {
+        let mut ts_model = TransitionSystemModel {
             name: model.name().into(),
             vars,
             state_predicates,
             transitions,
             specs,
+        };
+
+        // MD 2020-08-20: Moved "magic" from runner model helper to here.
+
+        // TODO:
+        // for now we dont do guard extraction because it appears we plan alot faster
+        // using just the invariants instead. but we need to make proper measurements
+
+        // each resource contains a supervisor defining its good states
+        let inits: Vec<Predicate> = ts_model.specs.iter()
+            .filter_map(|s| if s.name() == "supervisor" {
+                Some(s.invariant().clone())
+            } else {
+                None
+            })
+            .collect();
+
+        // we need to assume that we are in a state that adheres to the resources
+        let initial = Predicate::AND(inits);
+
+        if USE_GUARD_EXTRACTION {
+            let (new_guards, supervisor) = extract_guards(&ts_model, &initial);
+
+            // The specs are converted into guards + a global supervisor
+            ts_model.specs.clear();
+            ts_model.specs.push(Spec::new("global_supervisor", supervisor));
+
+            // TODO: right now its very cumbersome to update the original Model.
+            // but it would be nice if we could.
+            update_guards(&mut ts_model, &new_guards);
+        } else {
+            // we can refine all invariants instead of performing GE
+            let mut new_specs = Vec::new();
+            for s in &ts_model.specs {
+                println!("refining invariant {}", s.path());
+                let ri = refine_invariant(&ts_model, s.invariant());
+                let mut ns = s.clone();
+                ns.invariant = ri;
+                new_specs.push(ns);
+            }
+            ts_model.specs = new_specs;
         }
+
+        ts_model
     }
 
     pub fn from_op(model: &Model) -> Self {
@@ -158,7 +207,7 @@ impl TransitionSystemModel {
                 _ => None,
             }).collect()).unwrap_or(vec![]);
 
-        let mut transitions: Vec<Transition> =
+        let transitions: Vec<Transition> =
             model.find_item("operations",&[])
             .as_model()
             .map(|m| m
@@ -187,11 +236,4 @@ impl TransitionSystemModel {
     pub fn bad_state(&self, state: &SPState) -> bool {
         self.specs.iter().any(|s| !s.invariant().eval(state))
     }
-}
-
-#[derive(Debug, PartialEq, Clone, Default)]
-pub struct TransitionSystemResource {
-    pub path: SPPath,                          // the path of the resource
-    pub last_tick: Option<std::time::Instant>, // Last massage instant
-    pub last_echo: Option<SPState>,
 }
