+++ sp-runner/src/lib.rs
@@ -8,9 +8,6 @@ pub use sp_threaded_runner::*;
 mod modeling;
 pub use modeling::*;
 
-mod helpers;
-pub use helpers::*;
-
 mod planning;
 pub use planning::*;
